requirements.txt
openpyxl
pandas
# auto-py-to-exe
PyYAML